StudyBuddy/requirements.txt
google-generativeai==0.8.3
sentence-transformers==3.1.1
chromadb==0.5.18
fastapi==0.115.0
uvicorn==0.31.1
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.7
beautifulsoup4==4.12.3
numpy==2.1.1
pandas==2.2.3
streamlit==1.39.0
tiktoken==0.8.0